+++ R/slide.R
@@ -267,7 +267,7 @@ pslide_dfr <- function(
 #' my_diag <- function(...) {
 #'   data <- list(...)
 #'   fit <- lm(Count ~ Time, data = data)
-#'   tibble::tibble(fitted = fitted(fit), resid = residuals(fit))
+#'   tibble(fitted = fitted(fit), resid = residuals(fit))
 #' }
 #' pedestrian %>%
 #'   filter(Date <= as.Date("2015-01-31")) %>%